De-deplicate requests for the same token before the cache is populated
There is already an IDistributedCache to hold claims once they have been
fetched. But before the cache has been populated there is a window where
lots of threads can come in and make network requests for the same
token. I have added a keyed semaphore so if many threads want claims for
the same token, only one thread is able to make the request while the
others wait