https://github.com/IdentityModel/IdentityModel.AspNetCore.OAuth2Introspection/issues/130
Allow clients to control the how the cache key is generated from a token.
If a client can rely on IDistributedCache, a client can then have direct control over downstream caching functionality like Remove cache item.

Use Case:
The token that came in via the request has been mutated downstream, hence the claims that were generated from that token are no longer valid.  The claims in the cache vs the claims that the token is now pointing to are out of sync.

The downstream mutation flow must be able to remove the token from the cache so not to incur needless mutations for each new request for the duration of the cache.